requirements.base.txt
# ============================================================
# TMvault shared base requirements
# Union of all service/worker requirements — install once in
# tmvault-base:latest so every service image just COPYs code.
# ============================================================

# --- Web framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# --- Database ---
sqlalchemy==2.0.25
asyncpg==0.29.0
# Alembic for schema migrations. Replaces Base.metadata.create_all
# as the prod-safe schema-evolution path (see tm_backend/alembic/).
alembic>=1.13.0,<2

# --- HTTP client ---
# httpx + h2 to enable HTTP/2 multiplexing on the persistent Graph client.
# HTTP/2 reuses a single TCP connection across many concurrent requests
# (multiplexed streams), cutting handshake + TLS overhead to ~zero on the
# steady-state hot path. Graph supports it cleanly.
httpx==0.26.0
h2>=4.1.0,<5.0

# --- Message queue ---
aio-pika==9.4.0

# --- Pydantic ---
pydantic==2.5.3
pydantic-settings==2.1.0

# --- Auth / cryptography (needs gcc + libffi on ARM64) ---
cryptography==42.0.2
python-jose[cryptography]==3.3.0

# --- Cache ---
redis==5.0.1

# --- Scheduling ---
apscheduler==3.10.4

# --- Azure SDK ---
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-mgmt-storage==21.0.0

# --- S3 / SeaweedFS (on-prem backend) ---
# Pinned to minor versions that resolve cleanly with botocore 1.40.x.
aioboto3>=15.5.0,<16
aiobotocore>=2.25.0,<3
boto3>=1.40.0,<2

# --- Server-Sent Events (admin storage toggle stream) ---
sse-starlette==1.8.2

# --- Utilities ---
python-dotenv==1.0.0
